[@container] Properly invalidate container-relative units in gradients
This is not straightforward, because the actual unit conversion
happens outside of style resolution. It can also (in theory) happen
on a ComputedStyle which is not currently stored on an Element, which
is not compatible with our current behavior of marking ComputedStyles
as container-query-dependent during length conversion. (The we would
set the container-query-dependent flags on a ComputedStyle never seen
by the recalc machinery).

To fix this, we add functions for figuring out whether or not a generated
image depends on container-relative units, and use this information
ElementStyleResources-time (which is during style resolution) to mark
the ComputedStyle as container-query-dependent.

Moved a SetDependsOnContainerQueries call to happen directly from
CSSToLengthConversionData instead of an after-the-fact fixup in
StyleResolver::ResolveStyle. This is to avoid any problems with
length conversions that may happen *after* that check.

This also fixes the issue of fetching the container sizes eagerly if
they're not really needed (Issue 1324901).

Bug: 1223030
Change-Id: I78cd8ef571c0563b0d308d3801446ece7a0b1693
Fixed: 1313084, 1324901