[@container] Support container-relative units in gradients
Units in gradients are resolved outside of style resolution, using a
separate CSSToLengthConversionData object. (The main object being part
of StyleResolverState). At the time the gradients are produced, we don't
have enough context to pass the nearest container to the
CSSToLengthConversionData constructor.

This CL solves that by eagerly fetching container sizes during
ElementStyleResources::LoadPendingResources, and storing those sizes
on StyleGeneratedImage (which is created during style resolution).

Note that invalidation does not work yet.

Bug: 1223030, 1313084
Change-Id: I2373bca3dbadb9129975f62b49c15e5631a033e6